fix: clear the connectivity and inv warnings, and a dead function on Linux
workbench/mainwindow.cpp defined qtLibrariesPath() unconditionally but only
called it from the Q_OS_MACOS branch, so on Linux and Windows it was a dead
static function and -Werror rejected it. Guarded the definition to match its
only use rather than deleting it.

The five phase-based connectivity metrics share the same shape: a function level
"int i,j" where the later loops declare their own j. Only one loop per function
used the outer index, so it is now declared in each loop and dropped from the
joint declaration - 11 C4456 reports across the five files.

inv_dipole_fit_data had the same pattern for k and p. Note the outer k was used
by four separate loops, two of them far below the reported line, so removing the
declaration outright broke the build until each loop got its own; the fix is
per-loop scoping, not deletion. Care was also needed with "for (k = 0, do_it = 0
...)": scoping that whole init list would have shadowed do_it and silently
broken the check that follows it.

inv_ecd_set moves both dipole writers off FILE* to QFile. The bdip writer is
binary, so the record is written through QFile::write with the same size check;
the dip writer keeps printf style formatting because the column layout is read
back by other tools.

Verified: clang build 0 warnings, 181/181 ctest, and specifically the four test
frames that exercise these writers including the binary bdip round trip.

+++ src/libraries/inv/hpi/inv_hpi_fit.cpp
@@ -358,7 +358,7 @@ double InvHpiFit::objectTrans(const MatrixXd& matHeadCoil,
 Eigen::MatrixXd InvHpiFit::order(const std::vector<int>& vecOrder,
                               const Eigen::MatrixXd& matToOrder)
 {
-    const int iNumCoils = vecOrder.size();
+    const int iNumCoils = static_cast<int>(vecOrder.size());
     MatrixXd matToOrderTemp = matToOrder;
 
     for(int i = 0; i < iNumCoils; i++) {
@@ -372,7 +372,7 @@ Eigen::MatrixXd InvHpiFit::order(const std::vector<int>& vecOrder,
 QVector<int> InvHpiFit::order(const std::vector<int>& vecOrder,
                            const QVector<int>& vecToOrder)
 {
-    const int iNumCoils = vecOrder.size();
+    const int iNumCoils = static_cast<int>(vecOrder.size());
     QVector<int> vecToOrderTemp = vecToOrder;
 
     for(int i = 0; i < iNumCoils; i++) {
